Rename state methods from find_ to get_
For consistency with the rest of the library, rename the methods
prefixed with `find_` to `get_`.

The past logic was that items are "found", as they may or may not exist.
With get, the expectation is that it is _always_ there, i.e. over REST.
However, this is inconsistent, and "get"ting over REST can fail for
other reasons.