Paginate rest::get_guilds and add helper methods
Paginate the rest::get_guilds endpoint, which can be used to retrieve
the guilds after-or-before a certain guild's Id, with a limit of results
to return (100 is the max).

Example usage:

```rs
use serenity::client::rest::GuildPagination;

let target = GuildPagination::After(GuildId(81384788765712384));

let guilds = context.get_guilds(target, 50);
```

Additionally, add a method of `CurrentUser::guilds()`, to retrieve the
current user's guilds. A limit of 100 is set, as (most) users can not be
in more than 100 guilds.