Add a test suite for event deserialization
Add a test suite for most of the events that can be easily procuced,
and store them in static JSON files so the tests are ran on every build.

The recent update to using serde{,_derive} for deserialization was a
rough patch, as it was a large change which couldn't be fully tested at
the time. By adding JSON files which are deserialized, this will produce
a test suite that can be easily run on every new set of commits to
prevent any backwards incompatible changes with regards to
deserialization.