Move builtin framework impl to its own module
The framework is now moved in its entirity to the `framework` module,
with the `Framework` trait currently on its own and the builtin
implementation provided.

The builtin implementation has been renamed to "Standard".

Upgrade path:

Rename the `BuiltinFramework` import to `StandardFramework`. Instead of
importing builtin framework items from `serenity::framework`, import
them from `serenity::framework::standard`.

This is the beginning to #60. The root `framework` module (non-standard
implementation) will be built more by the time it's closed.