Fix User::has_role
Fix the return value of the function by properly checking whether the
user has the role in the given guild.

The function made the erraneous mistake of simply checking if the given
guild had the role by Id, not whether the _member in the given guild_
had the role by Id.